requirements.txt
asgiref==3.5.2
attrs==22.1.0
certifi==2022.9.24
charset-normalizer==2.1.1
coverage==6.5.0
Django==4.1.4
djangorestframework==3.14.0
exceptiongroup==1.0.4
idna==3.4
iniconfig==1.1.1
install==1.3.5
packaging==21.3
pluggy==1.0.0
pyparsing==3.0.9
pytest==7.2.0
pytest-django==4.5.2
pytz==2022.6
requests==2.28.1
sqlparse==0.4.3
tomli==2.0.1
urllib3==1.26.13
